Time: 181 ms (25.52%), Space: 99.2 MB (85.20%) - LeetHub

+++ 1838-frequency-of-the-most-frequent-element/1838-frequency-of-the-most-frequent-element.cpp
@@ -12,7 +12,7 @@ public:
             while(r < n) {
                 sum += nums[r];
                 int windowSize = r - l + 1;
-                long long windowSum = (long long)windowSize * nums[r];
+                long long windowSum = (long long)windowSize * (long long)nums[r];
                 if(k >= windowSum - sum) {
                     count++;
                     maxCount = max(maxCount, count);
